pyproject.toml
[tool.poetry]
name = "bindstuff"
version = "0.1.0"
description = "Protein binder design pipeline"
authors = ["No author specified"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
openmm = "^8.0"
numpy = "^2.1.0"
pandas = "^2.3.2"
matplotlib = "^3.10.5"
scipy = "^1.16.1"
prody = "^2.4.0"
jax = "0.4.20"
jaxlib = "0.4.20"
colabdesign = {git = "https://github.com/sokrypton/ColabDesign.git"}
biopython = "^1.81"
pytest = "^8.4.1"
absl-py = "^2.0.0"
dm-haiku = "0.0.12"
tensorflow = "^2.15.0"
dm-tree = "^0.1.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.11"
pytest-cov = "^5.0.0"
pylint = "^3.2.6"
deptry = "^0.17.0"
pandas-stubs = "^2.3.2.250827"
scipy-stubs = "^1.16.1.1"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
explicit_package_bases = true
exclude = ["pdbfixer/"]

# Handle missing stubs for scientific packages
[[tool.mypy.overrides]]
module = [
    "openmm.*",
    "Bio.*", 
    "prody.*",
    "colabdesign.*",
    "numpy.*",
    "jax.*",
    "matplotlib.*",
    "scipy.*",
    "pandas.*",
    "pdbfixer.*"
]
ignore_missing_imports = true
ignore_errors = true

[tool.poetry.extras]
pyrosetta = ["pyrosetta"]

[tool.deptry]
extend_exclude = ["alphafold", "pdbfixer", "bindstuff.py"]
per_rule_ignores = {DEP001 = ["pyrosetta"], DEP002 = ["absl-py", "dm-haiku", "tensorflow", "dm-tree", "urllib2", "cStringIO"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
